requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
geographiclib==2.0
geopy==2.4.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
pymongo==4.8.0
python-dotenv==1.0.1
shapely==2.0.5
Werkzeug==3.0.3
